requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
xlrd==2.0.1
